+++ src/main/glob.h
@@ -288,7 +288,7 @@ struct ENE_WRK {
 };
 
 typedef struct {
-	int dat_adr;
+	int64_t dat_adr;
 	u_char floor;
 	u_char load_area;
 	u_char now_room;
